src/api/requirements.txt
# FastAPI Backend Requirements
# Save this as requirements.txt in your backend directory
# Install with: pip install -r requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0